+++ CHANGELOG.md
@@ -28,6 +28,7 @@ Changes to this project are documented in this file. More detail and links can b
 - AUTOTOOL-2761: ([GitHub Issue 261](https://github.com/F5Networks/f5-declarative-onboarding/issues/261)): DO always enables dhcp on mgmt interface after post declaration
 - AUTOTOOL-2823: ([GitHub Issue 274](https://github.com/F5Networks/f5-declarative-onboarding/issues/274)): Unable to specify gw and target in Route class
 - AUTOTOOL-2780: ([GitHub Issue 268](https://github.com/F5Networks/f5-declarative-onboarding/issues/268)): DO 1.24.0 doesn't honor Remote Role Groups "remoteAccess": true setting. Previously working on 1.21.1
+- AUTOTOOL-2882: Unable to POST declaration with single RADIUS server
 
 ### Changed
 - AUTOTOOL-2544: Use a default of "UTC" for the timezone property of the NTP class